major-projects/news-trend-nlp/backend_django/requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
psycopg2-binary
dj-database-url
gunicorn
whitenoise
requests
python-dotenv
langchain
langchain-community
langchain-huggingface
